population: move in dwelltime analysis
