+++ lumicks/pylake/population/tests/data/exponential_data.npz
